Fixed drone orientation in tf_publisher and added use_gui flag in sim

+++ src/drone_mujoco/drone_mujoco/sim.py
@@ -1,5 +1,4 @@
 import mujoco
-import mujoco_viewer
 import numpy as np
 import rclpy
 from rclpy.node import Node
@@ -8,20 +7,29 @@ from std_msgs.msg import Float32MultiArray
 from geometry_msgs.msg import PointStamped
 from sensor_msgs.msg import Imu
 from ament_index_python.packages import get_package_share_directory
+import time
 
 class SimNode(Node):
-    def __init__(self):
+    def __init__(self, use_gui=False):
         super().__init__('sim_node')
+        self.use_gui = use_gui
 
         path = '/home/ws/src/drone_mujoco/model/scene.xml'
 
         self.model = mujoco.MjModel.from_xml_path(path)
         self.data = mujoco.MjData(self.model)
-        self.viewer = mujoco_viewer.MujocoViewer(self.model, self.data)
-
+        
+        if self.use_gui:
+            import mujoco_viewer
+            self.viewer = mujoco_viewer.MujocoViewer(self.model, self.data)
+        else:
+            self.viewer = None
+        
         self.model.opt.timestep = 0.005
-        self.create_timer(0.005, self.loop)
+        self.create_timer(self.model.opt.timestep, self.loop)
         self.running = True
+        
+        self.prev_time = 0.0
 
         self.motor_subscription = self.create_subscription(
             Float32MultiArray,
@@ -43,13 +51,12 @@ class SimNode(Node):
     def set_control(self, motor_commands):
         for i in range(4):
             self.data.ctrl[i] = np.clip(motor_commands[i], 0, 1)
-            # print(f"{motor_commands[i]}")
 
     def publish_state(self):
         x, y, z = self.data.qpos[0:3]
 
         gps_msg = PointStamped()
-        gps_msg.header.stamp = self.get_clock().now().to_msg() 
+        gps_msg.header.stamp = self.get_clock().now().to_msg()
         gps_msg.header.frame_id = "map"
         gps_msg.point.x, gps_msg.point.y, gps_msg.point.z = x, y, z
         self.gps_publisher.publish(gps_msg)
@@ -66,17 +73,24 @@ class SimNode(Node):
         self.imu_publisher.publish(imu_msg)
 
     def loop(self):
-        if self.viewer.is_alive:
-            mujoco.mj_step(self.model, self.data)
+        mujoco.mj_step(self.model, self.data)
+        
+        print(f"Iterations per second: {int(1.0 / (time.time() - self.prev_time))}")
+        self.prev_time = time.time()
+        
+        if self.viewer and self.viewer.is_alive:
             self.viewer.render()
-            self.publish_state()
-        else:
+        elif self.viewer and not self.viewer.is_alive:
             self.running = False
             self.destroy_node()
 
+        self.publish_state()
+        
+
 def main():
     rclpy.init()
-    node = SimNode()
+    use_gui = False  # CHANGE IF GUI NEEDED
+    node = SimNode(use_gui=use_gui)
     try:
         while rclpy.ok() and node.running:
             rclpy.spin_once(node)
@@ -84,6 +98,7 @@ def main():
         pass
     node.destroy_node()
     rclpy.shutdown()
+
 
 if __name__ == '__main__':
     main()
